refactor(tag-input): move hooks to floder
